requirements.txt
absl-py==1.4.0
aioboto3==11.3.0
aiobotocore==2.6.0
aiodns==3.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
aleph-alpha-client==2.17.0
altair==5.1.1
amadeus==9.0.0
anyio==4.0.0
arxiv==1.4.8
astunparse==1.6.3
async-timeout==4.0.3
atlassian-python-api==3.41.2
attr==0.3.2
attrs==23.1.0
audioread==3.0.0
Authlib==1.2.1
awadb==0.3.10
azure-ai-formrecognizer==3.3.0
azure-ai-vision==0.11.1b1
azure-cognitiveservices-speech==1.32.1
azure-common==1.1.28
azure-core==1.29.1
azure-cosmos==4.5.0
azure-identity==1.14.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.28.17
botocore==1.31.17
Brotli==1.1.0
build==1.0.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clarifai==9.8.0
clarifai-grpc==9.8.0
click==8.1.7
clickhouse-connect==0.5.25
cohere==4.24
cryptography==41.0.3
dataclasses-json==0.5.14
decorator==5.1.1
deeplake==3.6.25
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docarray==0.32.1
docker==6.1.3
duckduckgo-search==3.8.5
elastic-transport==8.4.0
elasticsearch==8.9.0
entrypoints==0.4
esprima==4.0.1
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastavro==1.8.2
feedparser==6.0.10
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.0
future==0.18.3
gast==0.4.0
geojson==2.5.0
gitdb==4.0.10
GitPython==3.1.34
google-api-core==2.11.1
google-api-python-client==2.70.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
gptcache==0.1.40
greenlet==2.0.2
grpcio==1.58.0
grpcio-tools==1.58.0
h11==0.14.0
h2==4.1.0
h5py==3.9.0
hnswlib==0.7.0
hpack==4.0.0
html2text==2020.1.16
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
humbug==0.3.2
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
jaraco.context==4.3.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jq==1.5.0
jsonlines==4.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
lancedb==0.1.16
langchain==0.0.285
langkit==0.0.19
langsmith==0.0.35
lark==1.1.7
lazy_loader==0.3
libclang==16.0.6
libdeeplake==0.0.60
librosa==0.10.1
llama-index==0.8.23.post1
llvmlite==0.40.1
loguru==0.7.1
lxml==4.9.3
lz4==4.3.2
manifest-ml==0.0.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marqo==1.3.1
marshmallow==3.20.1
mdurl==0.1.2
mmh3==3.1.0
momento==1.9.1
momento-wire-types==0.75.0
more-itertools==10.1.0
msal==1.23.0
msal-extensions==1.0.0
msgpack==1.0.5
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nebula3-python==3.4.0
neo4j==5.12.0
nest-asyncio==1.5.7
networkx==2.8.8
nlpcloud==1.1.44
nltk==3.8.1
nomic==1.1.14
numba==0.57.1
numcodecs==0.11.0
numexpr==2.8.5
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
O365==2.0.28
oauthlib==3.2.2
openai==0.28.0
openlm==0.0.5
opensearch-py==2.3.1
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.1
pandas==2.1.0
pathos==0.3.1
pdfminer.six==20221105
pexpect==4.8.0
pgvector==0.1.8
Pillow==9.5.0
pinecone-client==2.2.2
pinecone-text==0.4.2
platformdirs==3.10.0
pooch==1.7.0
portalocker==2.7.0
pox==0.3.3
ppft==1.7.6.7
protobuf==4.24.2
psycopg2-binary==2.9.7
ptyprocess==0.7.0
py==1.11.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
pylance==0.5.10
pymongo==4.5.0
Pympler==1.0.1
pyowm==3.3.0
pyparsing==3.1.1
pypdf==3.15.5
pyphen==0.14.0
pyproject_hooks==1.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-arango==7.6.2
python-dateutil==2.8.2
python-rapidjson==1.10
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pyvespa==0.33.0
PyYAML==6.0.1
qdrant-client==1.5.0
ratelimiter==1.2.0.post0
rdflib==6.3.2
redis==4.6.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.4.2
rpds-py==0.10.2
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
singlestoredb==0.7.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
SQLAlchemy==2.0.20
sqlitedict==2.1.0
sqlparams==5.1.0
streamlit==1.26.0
stringcase==1.2.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-text==2.13.0
termcolor==2.3.0
textstat==0.7.3
threadpoolctl==3.2.0
tigrisdb==1.0.0b6
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.3.3
tqdm==4.64.1
transformers==4.33.1
tritonclient==2.34.0
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
validators==0.21.0
watchdog==3.0.0
weaviate-client==3.23.2
websocket-client==1.6.2
Werkzeug==2.3.7
wget==3.2
whylabs-client==0.5.6
whylogs==1.2.6
whylogs-sketching==3.4.1.dev3
wikipedia==1.4.0
wolframalpha==5.0.0
wonderwords==2.2.0
wrapt==1.15.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
